tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.interpolate.outside.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.interpolate.outside</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.interpolate.outside</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  var g = ctx.createLinearGradient(25, 0, 75, 0);
  g.addColorStop(0.4, '#0f0');
  g.addColorStop(0.6, '#0f0');

  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 100, 50);
  _assertPixelApprox(canvas, 20,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 80,25, 0,255,0,255, 2);
  t.done();

});
</script>
